Refactored precondition view in ConfigWizard
