+++ UI/Dialogs/ConfigWizard.cs
@@ -9,6 +9,7 @@ using System.Windows.Forms;
 using MobiFlight;
 using MobiFlight.Base;
 using MobiFlight.UI.Forms;
+using MobiFlight.UI.Panels.Config;
 
 namespace MobiFlight.UI.Dialogs
 {
@@ -57,7 +58,22 @@ namespace MobiFlight.UI.Dialogs
 #else
             initWithoutArcazeCache();
 #endif
-            preparePreconditionPanel(dataSetConfig, filterGuid);            
+            preconditionPanel.preparePreconditionPanel(dataSetConfig, filterGuid);
+            initConfigRefDropDowns(dataSetConfig, filterGuid);            
+        }
+
+        private void initConfigRefDropDowns(DataSet dataSetConfig, string filterGuid)
+        {
+            _dataSetConfig = dataSetConfig;
+            DataRow[] rows = dataSetConfig.Tables["config"].Select("guid <> '" + filterGuid + "'");
+
+            // this filters the current config
+            DataView dv = new DataView(dataSetConfig.Tables["config"]);
+            dv.RowFilter = "guid <> '" + filterGuid + "'";
+
+            configRefPanel.SetConfigRefsDataView(dv, filterGuid);
+            displayLedDisplayPanel.SetConfigRefsDataView(dv, filterGuid);
+            //displayShiftRegisterPanel.SetConfigRefsDataView(dv, filterGuid);
         }
 
         public bool ConfigHasChanged()
@@ -77,6 +93,7 @@ namespace MobiFlight.UI.Dialogs
                 cfg.Preconditions.Add(new Precondition());
 
             originalConfig = cfg.Clone() as OutputConfigItem;
+
             InitializeComponent();
             comparisonSettingsPanel.Enabled = false;
             
@@ -89,11 +106,19 @@ namespace MobiFlight.UI.Dialogs
             tabControlFsuipc.SelectedIndex = lastTabActive;
 
             _initDisplayPanels();
-            _initPreconditionPanel();
+
+            // PRECONDITION PANEL
+            preconditionPanel.Init();
+            preconditionPanel.ErrorOnValidating += (sender, e) =>
+            {
+                tabControlFsuipc.SelectedTab = preconditionTabPage;
+            };
+
+            // FSUIPC PANEL
             fsuipcConfigPanel.setMode(true);
             fsuipcConfigPanel.syncFromConfig(cfg);
 
-            // displayLedDisplayComboBox.Items.Clear(); 
+            // SIMCONNECT SIMVARS PANEL
             simConnectPanel1.OnGetLVarListRequested += SimConnectPanel1_OnGetLVarListRequested;
             _execManager.GetSimConnectCache().LVarListUpdated += ConfigWizard_LVarListUpdated;
         }
@@ -109,37 +134,6 @@ namespace MobiFlight.UI.Dialogs
         private void ConfigWizard_LVarListUpdated(object sender, EventArgs e)
         {
             simConnectPanel1.LVars = (sender as List<String>);
-        }
-
-        private void _initPreconditionPanel()
-        {
-            preConditionTypeComboBox.Items.Clear();
-            List<ListItem> preconTypes = new List<ListItem>() {
-                new ListItem() { Value = "none",    Label = i18n._tr("LabelPrecondition_None") },
-                new ListItem() { Value = "config",  Label = i18n._tr("LabelPrecondition_ConfigItem") },
-                new ListItem() { Value = "pin",     Label = i18n._tr("LabelPrecondition_ArcazePin") }
-            };
-            preConditionTypeComboBox.DataSource = preconTypes;
-            preConditionTypeComboBox.DisplayMember = "Label";
-            preConditionTypeComboBox.ValueMember = "Value";
-            preConditionTypeComboBox.SelectedIndex = 0;
-
-            preconditionConfigComboBox.SelectedIndex = 0;
-            preconditionRefOperandComboBox.SelectedIndex = 0;
-
-            // init the pin-type config panel
-            List<ListItem> preconPinValues = new List<ListItem>() {
-                new ListItem() { Value = "0", Label = "Off" },
-                new ListItem() { Value = "1", Label = "On" },                
-            };
-
-            preconditionPinValueComboBox.DataSource = preconPinValues;
-            preconditionPinValueComboBox.DisplayMember = "Label";
-            preconditionPinValueComboBox.ValueMember = "Value";
-            preconditionPinValueComboBox.SelectedIndex = 0;
-
-            preconditionSettingsPanel.Enabled = false;
-            preconditionApplyButton.Visible = false;
         }
 
         protected void _initDisplayPanels () {
@@ -253,35 +247,6 @@ namespace MobiFlight.UI.Dialogs
             
         }        
 
-        private void preparePreconditionPanel(DataSet dataSetConfig, String filterGuid)
-        {
-            _dataSetConfig = dataSetConfig;
-            DataRow[] rows = dataSetConfig.Tables["config"].Select("guid <> '" + filterGuid +"'");         
-   
-            // this filters the current config
-            DataView dv = new DataView (dataSetConfig.Tables["config"]);
-            dv.RowFilter = "guid <> '" + filterGuid + "'";
-            preconditionConfigComboBox.DataSource = dv;
-            preconditionConfigComboBox.ValueMember = "guid";
-            preconditionConfigComboBox.DisplayMember = "description";
-
-            if (preconditionConfigComboBox.Items.Count == 0)
-            {
-                List<ListItem> preconTypes = new List<ListItem>() {
-                new ListItem() { Value = "none",    Label = i18n._tr("LabelPrecondition_None") },
-                new ListItem() { Value = "pin",     Label = i18n._tr("LabelPrecondition_ArcazePin") }
-                };
-                preConditionTypeComboBox.DataSource = preconTypes;
-                preConditionTypeComboBox.DisplayMember = "Label";
-                preConditionTypeComboBox.ValueMember = "Value";
-                preConditionTypeComboBox.SelectedIndex = 0;
-            }
-
-            configRefPanel.SetConfigRefsDataView(dv, filterGuid);
-            displayLedDisplayPanel.SetConfigRefsDataView(dv, filterGuid);
-            //displayShiftRegisterPanel.SetConfigRefsDataView(dv, filterGuid);
-        }
-
 #if ARCAZE
         /// <summary>
         /// sync the config wizard with the provided settings from arcaze cache such as available modules, ports, etc.
@@ -289,19 +254,17 @@ namespace MobiFlight.UI.Dialogs
         /// <param name="arcazeCache"></param>
         public void initWithArcazeCache (ArcazeCache arcazeCache)
         {
-            
+            List<String> PreconditionModuleList = new List<String>();
             // update the display box with
             // modules
             displayModuleNameComboBox.Items.Clear();
-            preconditionPinSerialComboBox.Items.Clear();
             displayModuleNameComboBox.Items.Add("-");
-            preconditionPinSerialComboBox.Items.Add("-");
-
+            
             foreach (IModuleInfo module in arcazeCache.getModuleInfo())
             {
                 arcazeFirmware[module.Serial] = module.Version;
                 displayModuleNameComboBox.Items.Add(module.Name + "/ " + module.Serial);
-                preconditionPinSerialComboBox.Items.Add(module.Name + "/ " + module.Serial);
+                PreconditionModuleList.Add(module.Name + "/ " + module.Port);
             }
 
 
@@ -314,7 +277,8 @@ namespace MobiFlight.UI.Dialogs
             }
 
             displayModuleNameComboBox.SelectedIndex = 0;
-            preconditionPinSerialComboBox.SelectedIndex = 0;            
+
+            preconditionPanel.SetModules(PreconditionModuleList);
         }
 #endif
 #if MOBIFLIGHT
@@ -327,9 +291,7 @@ namespace MobiFlight.UI.Dialogs
             // update the display box with
             // modules
             displayModuleNameComboBox.Items.Clear();
-            preconditionPinSerialComboBox.Items.Clear();
             displayModuleNameComboBox.Items.Add("-");
-            preconditionPinSerialComboBox.Items.Add("-");
 
             foreach (IModuleInfo module in _execManager.getMobiFlightModuleCache().getModuleInfo())
             {
@@ -340,7 +302,6 @@ namespace MobiFlight.UI.Dialogs
             }
 
             displayModuleNameComboBox.SelectedIndex = 0;
-            preconditionPinSerialComboBox.SelectedIndex = 0;
         }
 #endif
         /// <summary>
@@ -394,29 +355,8 @@ namespace MobiFlight.UI.Dialogs
 
             displayShiftRegisterPanel.SyncFromConfig(config);
 
-            preconditionListTreeView.Nodes.Clear();
-
-            foreach (Precondition p in config.Preconditions)
-            {
-                TreeNode tmpNode = new TreeNode();
-                tmpNode.Text = p.ToString();
-                tmpNode.Tag = p;
-                tmpNode.Checked = p.PreconditionActive;
-                try
-                {
-                    _updateNodeWithPrecondition(tmpNode, p);
-                    preconditionListTreeView.Nodes.Add(tmpNode);
-                }
-                catch (IndexOutOfRangeException e)
-                {
-                    Log.Instance.log("An orphaned precondition has been found", LogSeverity.Error);
-                    continue;
-                }                
-            }
-
-            overridePreconditionCheckBox.Checked = config.Preconditions.ExecuteOnFalse;
-            overridePreconditionTextBox.Text = config.Preconditions.FalseCaseValue;
-
+            preconditionPanel.syncFromOutputConfig(config);
+            
             return true;
         }
 
@@ -478,11 +418,7 @@ namespace MobiFlight.UI.Dialogs
 
             display_syncToConfig();
 
-            // sync the two properties that are not part of the preconditions list
-
-            config.Preconditions.ExecuteOnFalse = overridePreconditionCheckBox.Checked;
-            config.Preconditions.FalseCaseValue = overridePreconditionTextBox.Text;
-            
+            preconditionPanel.syncToOutputConfig(config);
 
             // sync panels
             displayPinPanel.syncToConfig(config);
@@ -1025,100 +961,6 @@ namespace MobiFlight.UI.Dialogs
             }
         }
 
-        private void preConditionTypeComboBox_SelectedIndexChanged(object sender, EventArgs e)
-        {
-            string selected = ((sender as ComboBox).SelectedItem as ListItem).Value;
-            preconditionSettingsGroupBox.Visible = selected != "none";
-            preconditionRuleConfigPanel.Visible = false;
-            preconditionRuleConfigPanel.Visible = selected == "config";
-            preconditionPinPanel.Visible = selected == "pin";
-        }
-
-        private void preconditionRuleConfigPanel_Validating(object sender, CancelEventArgs e)
-        {            
-        }
-        
-        private void preconditionRefValueTextBox_Validating(object sender, CancelEventArgs e)
-        {
-            if (!(preconditionRuleConfigPanel).Visible)
-            {
-                removeError(preconditionRefValueTextBox);
-                return;
-            }
-
-            if (preconditionRefValueTextBox.Text.Trim() == "")
-            {
-                e.Cancel = true;
-                tabControlFsuipc.SelectedTab = preconditionTabPage;
-                displayError(preconditionRefValueTextBox, i18n._tr("uiMessageConfigWizard_SelectComparison"));
-            }
-            else
-            {
-                removeError(preconditionRefValueTextBox);
-            }
-        }
-
-        private void preconditionPinSerialComboBox_Validating(object sender, CancelEventArgs e)
-        {
-            if (!(preconditionPinPanel).Visible)
-            {
-                removeError(preconditionRefValueTextBox);
-                return;
-            }
-
-            if (preconditionPinSerialComboBox.Items.Count > 1 && preconditionPinSerialComboBox.Text.Trim() == "-")
-            {
-                e.Cancel = true;
-                tabControlFsuipc.SelectedTab = preconditionTabPage;
-                preconditionPinSerialComboBox.Focus();
-                displayError(preconditionPinSerialComboBox, i18n._tr("uiMessageConfigWizard_SelectArcaze"));
-            }
-            else
-            {
-                removeError(preconditionPinSerialComboBox);
-            }
-
-        }
-
-        private void preconditionPinComboBox_Validating(object sender, CancelEventArgs e)
-        {
-            if (!(preconditionPinPanel).Visible)
-            {
-                removeError(preconditionPinComboBox);
-                return;
-            }
-
-            if (preconditionPinSerialComboBox.SelectedIndex > 0 && preconditionPinComboBox.SelectedIndex == -1)
-            {
-                e.Cancel = true;
-                tabControlFsuipc.SelectedTab = preconditionTabPage;
-                displayError(preconditionPinComboBox, i18n._tr("Please_select_a_pin"));
-            }
-            else
-            {
-                removeError(preconditionPinComboBox);
-            }
-        }
-
-        private void preconditionPortComboBox_Validating(object sender, CancelEventArgs e) {
-            if (!(preconditionPinPanel).Visible)
-            {
-                removeError(preconditionPortComboBox);
-                return;
-            }
-
-            if (preconditionPinSerialComboBox.SelectedIndex > 0 && preconditionPortComboBox.SelectedIndex == -1)
-            {
-                e.Cancel = true;
-                tabControlFsuipc.SelectedTab = preconditionTabPage;
-                displayError(preconditionPortComboBox, i18n._tr("Please_select_a_port"));
-            }
-            else
-            {
-                removeError(preconditionPortComboBox);
-            }
-        }
-
         private void displayPinTestButton_Click(object sender, EventArgs e)
         {
             _testModeStart();
@@ -1205,170 +1047,7 @@ namespace MobiFlight.UI.Dialogs
             }
         }
         */
-        
-        private void preconditionListTreeView_NodeMouseClick(object sender, TreeNodeMouseClickEventArgs e)
-        {
-            preconditionListTreeView.SelectedNode = e.Node;
-            //if (e.Button != System.Windows.Forms.MouseButtons.Left) return;
 
-            Precondition config = (e.Node.Tag as Precondition);
-            preConditionTypeComboBox.SelectedValue = config.PreconditionType;
-            preconditionSettingsPanel.Enabled = true;
-            preconditionApplyButton.Visible = true;
-            config.PreconditionActive = e.Node.Checked;
-
-            switch (config.PreconditionType)
-            {
-                case "config":
-                    try
-                    {                        
-                        preconditionConfigComboBox.SelectedValue = config.PreconditionRef;
-                    }
-                    catch (Exception exc)
-                    {
-                        // precondition could not be loaded
-                        Log.Instance.log("preconditionListTreeView_NodeMouseClick : Precondition could not be loaded, " + exc.Message, LogSeverity.Debug);
-                    }
-
-                    ComboBoxHelper.SetSelectedItem(preconditionRefOperandComboBox, config.PreconditionOperand);
-                    preconditionRefValueTextBox.Text = config.PreconditionValue;
-                    break;
-
-                case "pin":
-                    ArcazeIoBasic io = new ArcazeIoBasic(config.PreconditionPin);                    
-                    ComboBoxHelper.SetSelectedItemByPart(preconditionPinSerialComboBox, config.PreconditionSerial);
-                    preconditionPinValueComboBox.SelectedValue = config.PreconditionValue;
-                    preconditionPortComboBox.SelectedIndex = io.Port;
-                    preconditionPinComboBox.SelectedIndex = io.Pin;
-                    break;
-            }  
-
-            aNDToolStripMenuItem.Checked = config.PreconditionLogic == "and";
-            oRToolStripMenuItem.Checked = config.PreconditionLogic == "or";            
-        }
-
-        private void preconditionApplyButton_Click(object sender, EventArgs e)
-        {
-            // sync the selected node with the current settings from the panels
-            TreeNode selectedNode = preconditionListTreeView.SelectedNode;
-            if (selectedNode == null) return;
-
-            Precondition c = selectedNode.Tag as Precondition;
-            
-            c.PreconditionType = (preConditionTypeComboBox.SelectedItem as ListItem).Value;
-            switch (c.PreconditionType)
-            {
-                case "config":
-                    c.PreconditionRef = preconditionConfigComboBox.SelectedValue.ToString();
-                    c.PreconditionOperand = preconditionRefOperandComboBox.Text;
-                    c.PreconditionValue = preconditionRefValueTextBox.Text;
-                    c.PreconditionActive = true;
-                    break;
-                
-                case "pin":                    
-                    c.PreconditionSerial = preconditionPinSerialComboBox.Text;
-                    c.PreconditionValue = preconditionPinValueComboBox.SelectedValue.ToString();
-                    c.PreconditionPin = preconditionPortComboBox.Text + preconditionPinComboBox.Text;
-                    c.PreconditionActive = true;
-                    break;                    
-            }
-
-            _updateNodeWithPrecondition(selectedNode, c);
-        }    
-    
-        private void _updateNodeWithPrecondition (TreeNode node, Precondition p) 
-        {
-            String label = p.PreconditionLabel;
-            if (p.PreconditionType == "config")
-            {
-                String replaceString = "[unknown]";
-                if (_dataSetConfig != null)
-                {
-                    DataRow[] rows = _dataSetConfig.Tables["config"].Select("guid = '" + p.PreconditionRef + "'");
-                    if (rows.Count() == 0) throw new IndexOutOfRangeException(); // an orphaned entry has been found
-                    replaceString = rows[0]["description"] as String;
-                }
-                label = label.Replace("<Ref:" + p.PreconditionRef  + ">", replaceString);
-            }
-            else if (p.PreconditionType == "pin")
-            {
-                label = label.Replace("<Serial:" + p.PreconditionSerial + ">", p.PreconditionSerial.Split('/')[0]);
-            }
-            
-            label = label.Replace("<Logic:and>", " (AND)").Replace("<Logic:or>", " (OR)");
-            node.Checked = p.PreconditionActive;
-            node.Tag = p;
-            node.Text = label;
-            aNDToolStripMenuItem.Checked = p.PreconditionLogic == "and";
-            oRToolStripMenuItem.Checked = p.PreconditionLogic == "or";
-        }
-
-        private void addPreconditionToolStripMenuItem_Click(object sender, EventArgs e)
-        {
-            Precondition p = new Precondition();
-            TreeNode n = new TreeNode();
-            n.Tag = p;
-            config.Preconditions.Add(p);
-            preconditionListTreeView.Nodes.Add(n);
-            _updateNodeWithPrecondition(n, p);
-        }
-
-        private void andOrToolStripMenuItem_Click(object sender, EventArgs e)
-        {
-            TreeNode selectedNode = preconditionListTreeView.SelectedNode;
-            Precondition p = selectedNode.Tag as Precondition;
-            if ((sender as ToolStripMenuItem).Text == "AND")
-                p.PreconditionLogic = "and";
-            else
-                p.PreconditionLogic = "or";
-
-            _updateNodeWithPrecondition(selectedNode, p);            
-        }
-
-        private void removePreconditionToolStripMenuItem_Click(object sender, EventArgs e)
-        {
-            TreeNode selectedNode = preconditionListTreeView.SelectedNode;
-            Precondition p = selectedNode.Tag as Precondition;
-            config.Preconditions.Remove(p);
-            preconditionListTreeView.Nodes.Remove(selectedNode);
-        }
-
-        private void preconditionPinSerialComboBox_SelectedIndexChanged(object sender, EventArgs e)
-        {
-            // get the deviceinfo for the current arcaze
-            ComboBox cb = preconditionPinSerialComboBox;
-            String serial = SerialNumber.ExtractSerial(cb.SelectedItem.ToString());
-            // if (serial == "" && config.DisplaySerial != null) serial = ArcazeModuleSettings.ExtractSerial(config.DisplaySerial);
-
-            if (serial.IndexOf("SN") != 0)
-            {
-                preconditionPortComboBox.Items.Clear();
-                preconditionPinComboBox.Items.Clear();
-
-                List<ListItem> ports = new List<ListItem>();
-
-                foreach (String v in ArcazeModule.getPorts())
-                {
-                    ports.Add(new ListItem() { Label = v, Value = v });
-                    if (v == "B" || v == "E" || v == "H" || v == "K")
-                    {
-                        ports.Add(new ListItem() { Label = "-----", Value = "-----" });
-                    }
-
-                    if (v == "A" || v == "B")
-                    {
-                        preconditionPortComboBox.Items.Add(v);
-                    }
-                }
-
-                List<ListItem> pins = new List<ListItem>();
-                foreach (String v in ArcazeModule.getPins())
-                {
-                    pins.Add(new ListItem() { Label = v, Value = v });
-                    preconditionPinComboBox.Items.Add(v);
-                }
-            }
-        }
 
         private void interpolationCheckBox_CheckedChanged(object sender, EventArgs e)
         {
